DialogService/Dev: Fix missing cast in show_multiple_choice_dialog cButtons assignment

+++ src/Views.Win32/DialogService.cpp
@@ -42,7 +42,7 @@ size_t DialogService::show_multiple_choice_dialog(const std::vector<std::wstring
     .pszWindowTitle = title,
     .pszMainIcon = icon,
     .pszContent = str,
-    .cButtons = buttons.size(),
+    .cButtons = (UINT)buttons.size(),
     .pButtons = buttons.data(),
     };
